+++ src/rules/at-rule-empty-line-before/index.js
@@ -1,7 +1,10 @@
+import { isString } from "lodash"
+
 import {
   hasBlock,
   optionsHaveException,
   optionsHaveIgnored,
+  optionsHaveIgnoredAtRule,
   report,
   ruleMessages,
   validateOptions,
@@ -24,6 +27,7 @@ export default function (expectation, options) {
       possible: {
         except: [ "blockless-group", "first-nested", "all-nested" ],
         ignore: [ "blockless-group", "after-comment", "all-nested" ],
+        ignoreAtRules: [isString],
       },
       optional: true,
     })
@@ -33,6 +37,9 @@ export default function (expectation, options) {
 
       // Ignore the first node
       if (atRule === root.first) { return }
+
+      // Return early if at-rule is to be ignored
+      if (optionsHaveIgnoredAtRule(options, atRule)) { return }
 
       // Optionally ignore the expectation if the node is blockless
       if (optionsHaveIgnored(options, "blockless-group") && !hasBlock(atRule)) { return }